+++ src/routes/AuthLayoutRouter/components/SiteTour/index.js
@@ -1,0 +1,3 @@
+import component from './SiteTour'
+
+export default component