+++ src/routes/GroupDetail/GroupDetail.js
@@ -326,14 +326,16 @@ export function SuggestedSkills ({ addSkill, currentUser, group, removeSkill }) 
     }
   }
 
-  return <div styleName='g.joinQuestion'>
-    <h4>Which of the following skills &amp; interests are relevant to you?</h4>
-    <div styleName='g.skillPills'>
-      <Pillbox
-        pills={pills}
-        handleClick={handleClick}
-        editable={false}
-      />
+  return (
+    <div styleName='g.joinQuestion'>
+      <h4>Which of the following skills & interests are relevant to you?</h4>
+      <div styleName='g.skillPills'>
+        <Pillbox
+          pills={pills}
+          handleClick={handleClick}
+          editable={false}
+        />
+      </div>
     </div>
-  </div>
+  )
 }