+++ src/routes/GroupWelcomeModal/index.js
@@ -1,3 +1,3 @@
-import GroupWelcomeModal from './GroupWelcomeModal'
-import connector from './GroupWelcomeModal.connector'
-export default connector(GroupWelcomeModal)
+import component from './GroupWelcomeModal'
+
+export default component