+++ src/routes/MemberSidebar/index.js
@@ -1,4 +1,0 @@
-import component from './MemberSidebar'
-import connector from './MemberSidebar.connector'
-
-export default connector(component)