+++ src/routes/Messages/PeopleSelector/PeopleSelector.test.js
@@ -14,7 +14,7 @@ const defaultProps = {
   removePerson: () => {},
   changeQuerystringParam: () => {},
   selectedPeople: [],
-  onCloseURL: '',
+  onCloseLocation: { pathname: '' },
   peopleSelectorOpen: true
 }