+++ src/routes/NonAuthLayout/Login/Login.connector.test.js
@@ -1,8 +1,0 @@
-import { mapStateToProps } from './Login.connector'
-describe('Login', () => {
-  it('returns the right keys', () => {
-    expect(mapStateToProps({}, { location: { search: '' } }).hasOwnProperty('error')).toBeTruthy()
-    expect(mapStateToProps({}, { location: { search: '' } }).hasOwnProperty('downloadAppUrl')).toBeTruthy()
-    expect(mapStateToProps({}, { location: { search: '' } }).hasOwnProperty('returnToURL')).toBeTruthy()
-  })
-})