+++ src/routes/NonAuthLayout/index.js
@@ -1,3 +1,0 @@
-import component from './NonAuthLayout'
-
-export default component