+++ src/routes/NonAuthLayoutRouter/Signup/FinishRegistration/index.js
@@ -1,0 +1,3 @@
+import component from './FinishRegistration'
+
+export default component