+++ src/store/reducers/locationHistory.js
@@ -10,7 +10,7 @@ export default (state = initialState, action) => {
     case LOCATION_CHANGE:
       return {
         previousLocation: state.currentLocation,
-        currentLocation: action.payload.location.pathname
+        currentLocation: action.payload.location
       }
     default:
       return state