+++ src/store/selectors/getPreviousLocation.js
@@ -1,3 +1,3 @@
 import { get } from 'lodash/fp'
 
-export default (state) => get('locationHistory.previousLocation', state) || '/'
+export default (state) => get('locationHistory.previousLocation', state) || { pathname: '/' }