+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to Hylo Evo (the Hylo front-end) will be documented in this 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.6.0] - 2022-04-28
+
+### Changed
+- Major refactor of authorized vs non-authorized routing, see `RootRouter`, `AuthLayoutRouter`, and `NonAuthLayoutRouter`
+- React Testing Library (RTL) setup simplified and notated
+
+### Added
+- Changes to authorization and authentication flow to accomodate related API changes
+- Adds new RTL-based tests for new and existing components
+- You can now click on the map to create posts at that location. Either click and hold for a couple seconds, or click the (+) button and then click on the map to add a post at that lat,lng on the map.
+
 ## [3.5.2] - 2022-03-28
 
 ### Fixed